% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dxDat}
\alias{dxDat}
\title{Sample Data of Diagnosis Code}
\format{A data frame with 1000 rows and 11 variables.}
\usage{
data(dxDat)
}
\description{
A data frame of diagnosis code with variates named
\code{drg} and \code{dx1}, \code{dx2}, ..., \code{dx10},
where
\itemize{
    \item \code{drg}: drg code for comorbidity;
    \item \code{dx1-dx10}: icd-9 code for 10 diagnosises.
}
}
\examples{
data(dxDat)
drg <- dxDat$drg
icd <- dxDat[, - 1L]
}
\keyword{datasets}
